Inline create-form-router
