+++ controllers/apply/views/error.njk
@@ -1,0 +1,17 @@
+{% extends "layouts/main.njk" %}
+{% block title %}Error | {{ form.title }} | {% endblock %}
+{% set bodyClass = 'has-static-header' %}
+
+{% block content %}
+    <div class="content-box u-inner-wide-only accent--pink a--border-top">
+        <div class="u-constrained-content s-prose u-margin-bottom">
+            <div class="form-message">
+                <h1 class="t1 form-message__title">{{ stepConfig.title }}</h1>
+            </div>
+
+            {{ stepConfig.message | safe }}
+
+            <p><a class="btn" href="{{returnUrl}}">Back</a></p>
+        </div>
+    </div>
+{% endblock %}
